Add mockery as dev dependency
The project is complex enough that, for example, to do a simple round-trip
test (get/setobject on XmlDeserializationVisitor) required me to chase down at
least third-level dependencies for the GraphNavigator.  That's rather pointless
when I don't care about testing the navigator, just the deserialisation visitor.

This might also make testing other parts of the codebase easier, by reducing
the need to chase down and instantiate multiple levels of dependencies.