Handle ObjectConstructor returning NULL
The DeserializationGraphNavigator calls the construct method on an
instance of ObjectConstructorInterface. According to the return type
hint of that method, it is allowed to return NULL. However, when this
occurs, the navigator will pass NULL to the startVisitingObject method
of the visitor. startVisitingObject expects an object to be passed in,
thus causing the serializer to crash with a type error.

We can prevent this error by calling the `visitNull` method if we detect
that the ObjectConstructor has returned NULL instead of an object.