Use array_values
When `filterNullFromTypes` receives the following `[ 0 => IdentifierTypeNode("null"), 1 => IdentifierTypeNode("string") ]` types, it will map the one that is called `"null"` with `null` like this: `[ 0 => null, 1 => IdentifierTypeNode("string") ]`.

It then passes the array to `array_filter` that removes all the `null` values, leaving the following array as a result: `[1 => IdentifierTypeNode("string") ]`.

The problem is that `array_filter` keeps the keys intact, while the code that uses this doesn't expect it.

By using `array_values` we get a clean array again: `[0 => IdentifierTypeNode("string") ]`.